+++ pyplant/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+FLASK_SECRET_KEY = os.getenv('FLASK_SECRET_KEY')
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = FLASK_SECRET_KEY
+app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///site.db" # relative path
+db = SQLAlchemy(app)
+
+from pyplant import routes
